Generated by Spark: After searching prduces the results please allow an option for customers to edit and refine the reference, in addition, the beginning menu should offer multiple referencing styles such as "Havard referencing" etc...
Furhter more I just tried it and I don't think the smart search is accurate, it didn't show the reference I was looking for, is there any way to make the search even more accurate?

Finally, when a reference has been made and added to the bibliography, there should also be an option to allow the user to copy the in-text version of the reference also.